requirements.txt
python==3.8.0
Django==3.0
virtualenv==16.7.8
djangorestframework==3.11.0
psycopg2==2.8.4